falta carrito

+++ js/carrito.js
@@ -1,11 +1,14 @@
+const serviciosEnCarrito = JSON.parse(localStorage.getItem("servicios-en-carrito"));
 
 const contenedorCarritoVacio = document.querySelector("#carrito-vacio");
 const contenedorCarritoServicios = document.querySelector("#carrito-servicio");
 const contenedorCarritoAcciones = document.querySelector("#carrito-acciones");
 const contenedorCarritoComprado = document.querySelector("#carrito-comprado");
-let botonesEliminar = document.querySelectorAll("#carrito-servicio-eliminar");
-const botonVaciar = document.querySelector("#carrito-acciones-vaciar");
 const contenedorTotal = document.querySelector("#total");
+const botonVaciar = document.querySelector("#carrito-acciones-vaciar");
+const botonComprar = document.querySelector("#carrito-acciones-comprar");
+let botonesEliminar = document.querySelectorAll(".carrito-sercicio-eliminar");
+
 
 function cargarServiciosCarrito() {
   if (serviciosEnCarrito && serviciosEnCarrito.length > 0) {
@@ -16,31 +19,37 @@ function cargarServiciosCarrito() {
 
     contenedorCarritoServicios.innerHTML = "";
 
-    serviciosEnCarrito.forEach((servicios) => {
+    serviciosEnCarrito.forEach(servicio => {
+
       const div = document.createElement("div");
       div.classList.add("carrito-servicio");
+
       div.innerHTML = `
-        <img class="carrito-servicio-imagen" src="${servicios.imagen}" alt="${servicios.titulo}">
+        <img class="carrito-servicio-imagen" src="${servicio.imagen}" alt="${servicio.titulo}">
         <div class="carrito-servicio-titulo">
           <small>Título</small>
-          <h3>${servicios.titulo}</h3>
+          <h3>${servicio.titulo}</h3>
         </div>
         <div class="carrito-servicio-cantidad">
           <small>Cantidad</small>
-          <p>${servicios.cantidad}</p>
+          <p>${servicio.cantidad}</p>
         </div>
         <div class="carrito-servicio-precio">
           <small>Precio</small>
-          <p>$${servicios.precio}</p>
+          <p>$${servicio.precio}</p>
         </div>
         <div class="carrito-servicio-subtotal">
           <small>Subtotal</small>
-          <p>$${servicios.precio * servicios.cantidad}</p>
+          <p>$${servicio.precio * servicio.cantidad}</p>
         </div>
-        <button class="carrito-servicio-eliminar" id="${servicios.id}"><i class="bi bi-trash-fill"></i></button>
+        <button class="carrito-servicio-eliminar" id="${servicio.id}"><i class="bi bi-trash-fill"></i></button>
       `;
       contenedorCarritoServicios.append(div);
     });
+
+    actualizarBotonesEliminar();
+    actualizarTotal();
+
   } else {
     contenedorCarritoVacio.classList.remove("disabled");
     contenedorCarritoServicios.classList.add("disabled");
@@ -49,36 +58,50 @@ function cargarServiciosCarrito() {
   }
   actualizarBotonesEliminar();
   actualizarTotal();
-}
+}  
+
+cargarServiciosCarrito();
 
 function actualizarBotonesEliminar() {
-  botonEliminar = document.querySelectorAll(".carrito-servicios-eliminar");
-  botonEliminar.forEach((boton) => {
-    boton.addEventListener("click", eliminarDelCarrito);
+  const botonesEliminar = document.querySelectorAll(".carrito-servicio-eliminar");
+
+  botonesEliminar.forEach(boton => {
+      boton.addEventListener("click", eliminarDelCarrito);
   });
 }
 
 function eliminarDelCarrito(e) {
   const idBoton = e.currentTarget.id;
-  const index = serviciosEnCarrito.findIndex((servicio) => servicio.id === idBoton);
+  const index = serviciosEnCarrito.findIndex(servicio => servicio.id === idBoton);
+
   serviciosEnCarrito.splice(index, 1);
   cargarServiciosCarrito();
   localStorage.setItem("servicios-en-carrito", JSON.stringify(serviciosEnCarrito));
 }
 
+
+botonVaciar.addEventListener("click", vaciarCarrito);
 function vaciarCarrito() {
-  serviciosEnCarrito.length = 0;
-  localStorage.setItem("servicios-en-carrito", JSON.stringify(serviciosEnCarrito));
-  cargarServiciosCarrito();
-}
+
+ if (result.isConfirmed) {
+          serviciosEnCarrito.length = 0;
+            localStorage.setItem("servicios-en-carrito", JSON.stringify(serviciosEnCarrito));
+            cargarServiciosCarrito();
+        }
+      }
+
 
 function actualizarTotal() {
-  const totalCalculo = serviciosEnCarrito.reduce((acc, servicio) => acc + servicio.precio * servicio.cantidad, 0);
+  const totalCalculo = serviciosEnCarrito.reduce((acc, servicio) => acc + (servicio.precio * servicio.cantidad), 0);
   contenedorTotal.innerText = `$${totalCalculo}`;
 }
 
-const botonComprar = document.querySelector("#boton-comprar");
-botonComprar.addEventListener("click", comprarCarrito);
+
+
+
+// COMPRAR
+
+  botonComprar.addEventListener("click", comprarCarrito);
 
 function comprarCarrito() {
   serviciosEnCarrito.length = 0;
@@ -89,5 +112,3 @@ function comprarCarrito() {
   contenedorCarritoAcciones.classList.add("disabled");
   contenedorCarritoComprado.classList.remove("disabled");
 }
-
-cargarServiciosCarrito();
